cli/requirements.txt
web3==6.15.1
rich==13.7.0
click==8.1.7
python-dotenv==1.0.1
requests==2.31.0
